Se modificaron los ID a long - Se implementó el método deletePersona(id)

+++ src/main/java/com/firstcrud/demo/model/Estado.java
@@ -14,7 +14,7 @@ public class Estado {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String nombre;
 
     @ManyToOne
@@ -29,15 +29,15 @@ public class Estado {
         this.pais = pais;
     }
 
-    public Estado(int id) {
+    public Estado(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
