+++ src/main/java/com/firstcrud/demo/model/Pais.java
@@ -12,7 +12,7 @@ public class Pais {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String nombre;
 
     public Pais() {
@@ -22,15 +22,15 @@ public class Pais {
         this.nombre = nombre;
     }
 
-    public Pais(int id) {
+    public Pais(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
